# sdk/openapi/studio.json
{
  "openapi": "3.0.1",
  "components": {
    "schemas": {
...
      "Input_97df4a50a602fc26f1ee6311a5f68da1d96e2143f76deaf46ec88c673e7b5125": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Output_20d14e6c9e21afc4a0245cde30dbea83a572b7611ecb2a2d7721e85b2f520260": {
        "type": "object",
        "properties": {
          "entity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Output_8f9918cf758d7828437221562274ce1aadfefc572741f4475e8f668456c57771"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ]
      },
...
      "Output_3c2625259fdd4db44d2677d12aa2f8cd153ff19765e0207f189c5b0699b54243": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
...
      "Output_5ed634d1c74bb010a08c26559915f494ded8d053720f174c362df0724cd84c89": {
        "type": "object",
        "properties": {
          "entity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Output_819a96401e1c94505523cbabb8e79d80a54fa448e4025d35e66df55d830f8d2f"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ]
      },
...
      "Output_819a96401e1c94505523cbabb8e79d80a54fa448e4025d35e66df55d830f8d2f": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
...
      "Output_8f9918cf758d7828437221562274ce1aadfefc572741f4475e8f668456c57771": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
...
      "Output_a6d9fb3f70e6b7863e0c88327d238407dbfc8a53a8fbbcb08c080358f1491ec6": {
        "type": "object",
        "properties": {
          "entity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Output_ca207e9470562713039abace94a8b79c331512fc095ce1e937d8c9887565342b"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ]
      },
...
      "Output_b54579a1cbe27afaa9912714bdb85378479c70d6bd7a9c3058c8582378ddcd59": {
        "type": "object",
        "properties": {
          "entity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Output_3c2625259fdd4db44d2677d12aa2f8cd153ff19765e0207f189c5b0699b54243"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ]
      },
      "Output_bd7a1b00fadef6f5e471ef21cd4fe9df0f1e3f4b9cabd09814c00b321086588a": {
        "type": "object",
        "properties": {
          "canManage": {
            "type": "boolean"
          },
          "choices": {
            "$ref": "#/components/schemas/Output_c97b7b59c0c99ea53e6472f7e79baee9949576f0200174c238c8b514fcc51b69"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "choices",
          "canManage",
          "source"
        ]
      },
...
      "Output_c97b7b59c0c99ea53e6472f7e79baee9949576f0200174c238c8b514fcc51b69": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Output_b54579a1cbe27afaa9912714bdb85378479c70d6bd7a9c3058c8582378ddcd59"
            },
            "nullable": true
          },
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "exposure": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Output_5ed634d1c74bb010a08c26559915f494ded8d053720f174c362df0724cd84c89"
            },
            "nullable": true
          },
          "role": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Output_a6d9fb3f70e6b7863e0c88327d238407dbfc8a53a8fbbcb08c080358f1491ec6"
            },
            "nullable": true
          },
          "subject": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Output_20d14e6c9e21afc4a0245cde30dbea83a572b7611ecb2a2d7721e85b2f520260"
            },
            "nullable": true
          }
        },
        "required": [
          "subject",
          "role",
          "exposure",
          "entity",
          "entityTypes"
        ]
      },
      "Output_ca207e9470562713039abace94a8b79c331512fc095ce1e937d8c9887565342b": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "Datly Studio SDK",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/studio/sdk/access.context": {
      "post": {
        "operationId": "POST:/v1/studio/sdk/access.context",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Input_97df4a50a602fc26f1ee6311a5f68da1d96e2143f76deaf46ec88c673e7b5125"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Output_bd7a1b00fadef6f5e471ef21cd4fe9df0f1e3f4b9cabd09814c00b321086588a"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
...
  }
}